check directory before writing a file (#1327)
(cherry picked from commit 3965a3a52b5ef1e33d8d49bd187805daf3d95c4c)

+++ insights/client/utilities.py
@@ -95,6 +95,8 @@ def write_to_disk(filename, delete=False, content=get_time()):
     """
     Write filename out to disk
     """
+    if not os.path.exists(os.path.dirname(filename)):
+        return
     if delete:
         if os.path.lexists(filename):
             os.remove(filename)
